Move the execute method from the AthenaClient client class to the Athena facade.
